2.4.112.4
Added DummyStateMap
Added WeightedEnum

Changed LoadingUtils to print stack trace

Fixed Item Model method
Fixed TileEntity syncing
Fixed server join exception (attempt)

Removed the ASM config file

Signed jar

Code cleanup

+++ src/main/java/DummyCore/Client/GuiElement.java
@@ -1,6 +1,5 @@
 package DummyCore.Client;
 
-import DummyCore.Utils.TessellatorWrapper;
 import net.minecraft.client.gui.FontRenderer;
 import net.minecraft.client.renderer.BufferBuilder;
 import net.minecraft.client.renderer.GlStateManager;
@@ -41,7 +40,7 @@ public abstract class GuiElement {
 	 */
 	public abstract int getY();
 
-	//Generic GUI functions 
+	//Generic GUI functions
 
 	/**
 	 * Draws a 1 pixel wide horizontal line.
@@ -85,10 +84,10 @@ public abstract class GuiElement {
 			bottom = j;
 		}
 
-		float f3 = (float)(color >> 24 & 255) / 255.0F;
-		float f = (float)(color >> 16 & 255) / 255.0F;
-		float f1 = (float)(color >> 8 & 255) / 255.0F;
-		float f2 = (float)(color & 255) / 255.0F;
+		float f3 = (color >> 24 & 255) / 255.0F;
+		float f = (color >> 16 & 255) / 255.0F;
+		float f1 = (color >> 8 & 255) / 255.0F;
+		float f2 = (color & 255) / 255.0F;
 		Tessellator tessellator = Tessellator.getInstance();
 		BufferBuilder bufferbuilder = tessellator.getBuffer();
 		GlStateManager.enableBlend();
@@ -96,10 +95,10 @@ public abstract class GuiElement {
 		GlStateManager.tryBlendFuncSeparate(GlStateManager.SourceFactor.SRC_ALPHA, GlStateManager.DestFactor.ONE_MINUS_SRC_ALPHA, GlStateManager.SourceFactor.ONE, GlStateManager.DestFactor.ZERO);
 		GlStateManager.color(f, f1, f2, f3);
 		bufferbuilder.begin(7, DefaultVertexFormats.POSITION);
-		bufferbuilder.pos((double)left, (double)bottom, 0.0D).endVertex();
-		bufferbuilder.pos((double)right, (double)bottom, 0.0D).endVertex();
-		bufferbuilder.pos((double)right, (double)top, 0.0D).endVertex();
-		bufferbuilder.pos((double)left, (double)top, 0.0D).endVertex();
+		bufferbuilder.pos(left, bottom, 0.0D).endVertex();
+		bufferbuilder.pos(right, bottom, 0.0D).endVertex();
+		bufferbuilder.pos(right, top, 0.0D).endVertex();
+		bufferbuilder.pos(left, top, 0.0D).endVertex();
 		tessellator.draw();
 		GlStateManager.enableTexture2D();
 		GlStateManager.disableBlend();
@@ -109,14 +108,14 @@ public abstract class GuiElement {
 	 * Draws a rectangle with a vertical gradient between the specified colors (ARGB format).
 	 */
 	public void drawGradientRect(int left, int top, int right, int bottom, int startColor, int endColor) {
-		float f = (float)(startColor >> 24 & 255) / 255.0F;
-		float f1 = (float)(startColor >> 16 & 255) / 255.0F;
-		float f2 = (float)(startColor >> 8 & 255) / 255.0F;
-		float f3 = (float)(startColor & 255) / 255.0F;
-		float f4 = (float)(endColor >> 24 & 255) / 255.0F;
-		float f5 = (float)(endColor >> 16 & 255) / 255.0F;
-		float f6 = (float)(endColor >> 8 & 255) / 255.0F;
-		float f7 = (float)(endColor & 255) / 255.0F;
+		float f = (startColor >> 24 & 255) / 255.0F;
+		float f1 = (startColor >> 16 & 255) / 255.0F;
+		float f2 = (startColor >> 8 & 255) / 255.0F;
+		float f3 = (startColor & 255) / 255.0F;
+		float f4 = (endColor >> 24 & 255) / 255.0F;
+		float f5 = (endColor >> 16 & 255) / 255.0F;
+		float f6 = (endColor >> 8 & 255) / 255.0F;
+		float f7 = (endColor & 255) / 255.0F;
 		GlStateManager.disableTexture2D();
 		GlStateManager.enableBlend();
 		GlStateManager.disableAlpha();
@@ -125,10 +124,10 @@ public abstract class GuiElement {
 		Tessellator tessellator = Tessellator.getInstance();
 		BufferBuilder bufferbuilder = tessellator.getBuffer();
 		bufferbuilder.begin(7, DefaultVertexFormats.POSITION_COLOR);
-		bufferbuilder.pos((double)right, (double)top, (double)this.zLevel).color(f1, f2, f3, f).endVertex();
-		bufferbuilder.pos((double)left, (double)top, (double)this.zLevel).color(f1, f2, f3, f).endVertex();
-		bufferbuilder.pos((double)left, (double)bottom, (double)this.zLevel).color(f5, f6, f7, f4).endVertex();
-		bufferbuilder.pos((double)right, (double)bottom, (double)this.zLevel).color(f5, f6, f7, f4).endVertex();
+		bufferbuilder.pos(right, top, this.zLevel).color(f1, f2, f3, f).endVertex();
+		bufferbuilder.pos(left, top, this.zLevel).color(f1, f2, f3, f).endVertex();
+		bufferbuilder.pos(left, bottom, this.zLevel).color(f5, f6, f7, f4).endVertex();
+		bufferbuilder.pos(right, bottom, this.zLevel).color(f5, f6, f7, f4).endVertex();
 		tessellator.draw();
 		GlStateManager.shadeModel(7424);
 		GlStateManager.disableBlend();
@@ -140,14 +139,14 @@ public abstract class GuiElement {
 	 * Renders the specified text to the screen, center-aligned.
 	 */
 	public void drawCenteredString(FontRenderer fontRendererIn, String text, int x, int y, int color) {
-		fontRendererIn.drawStringWithShadow(text, (float)(x - fontRendererIn.getStringWidth(text) / 2), (float)y, color);
+		fontRendererIn.drawStringWithShadow(text, x - fontRendererIn.getStringWidth(text) / 2, y, color);
 	}
 
 	/**
 	 * Renders the specified text to the screen.
 	 */
 	public void drawString(FontRenderer fontRendererIn, String text, int x, int y, int color) {
-		fontRendererIn.drawStringWithShadow(text, (float)x, (float)y, color);
+		fontRendererIn.drawStringWithShadow(text, x, y, color);
 	}
 
 	/**
@@ -159,10 +158,10 @@ public abstract class GuiElement {
 		Tessellator tessellator = Tessellator.getInstance();
 		BufferBuilder bufferbuilder = tessellator.getBuffer();
 		bufferbuilder.begin(7, DefaultVertexFormats.POSITION_TEX);
-		bufferbuilder.pos((double)(x + 0), (double)(y + height), (double)this.zLevel).tex((double)((float)(textureX + 0) * 0.00390625F), (double)((float)(textureY + height) * 0.00390625F)).endVertex();
-		bufferbuilder.pos((double)(x + width), (double)(y + height), (double)this.zLevel).tex((double)((float)(textureX + width) * 0.00390625F), (double)((float)(textureY + height) * 0.00390625F)).endVertex();
-		bufferbuilder.pos((double)(x + width), (double)(y + 0), (double)this.zLevel).tex((double)((float)(textureX + width) * 0.00390625F), (double)((float)(textureY + 0) * 0.00390625F)).endVertex();
-		bufferbuilder.pos((double)(x + 0), (double)(y + 0), (double)this.zLevel).tex((double)((float)(textureX + 0) * 0.00390625F), (double)((float)(textureY + 0) * 0.00390625F)).endVertex();
+		bufferbuilder.pos(x + 0, y + height, this.zLevel).tex((textureX + 0) * 0.00390625F, (textureY + height) * 0.00390625F).endVertex();
+		bufferbuilder.pos(x + width, y + height, this.zLevel).tex((textureX + width) * 0.00390625F, (textureY + height) * 0.00390625F).endVertex();
+		bufferbuilder.pos(x + width, y + 0, this.zLevel).tex((textureX + width) * 0.00390625F, (textureY + 0) * 0.00390625F).endVertex();
+		bufferbuilder.pos(x + 0, y + 0, this.zLevel).tex((textureX + 0) * 0.00390625F, (textureY + 0) * 0.00390625F).endVertex();
 		tessellator.draw();
 	}
 
@@ -175,10 +174,10 @@ public abstract class GuiElement {
 		Tessellator tessellator = Tessellator.getInstance();
 		BufferBuilder bufferbuilder = tessellator.getBuffer();
 		bufferbuilder.begin(7, DefaultVertexFormats.POSITION_TEX);
-		bufferbuilder.pos((double)(xCoord + 0.0F), (double)(yCoord + (float)maxV), (double)this.zLevel).tex((double)((float)(minU + 0) * 0.00390625F), (double)((float)(minV + maxV) * 0.00390625F)).endVertex();
-		bufferbuilder.pos((double)(xCoord + (float)maxU), (double)(yCoord + (float)maxV), (double)this.zLevel).tex((double)((float)(minU + maxU) * 0.00390625F), (double)((float)(minV + maxV) * 0.00390625F)).endVertex();
-		bufferbuilder.pos((double)(xCoord + (float)maxU), (double)(yCoord + 0.0F), (double)this.zLevel).tex((double)((float)(minU + maxU) * 0.00390625F), (double)((float)(minV + 0) * 0.00390625F)).endVertex();
-		bufferbuilder.pos((double)(xCoord + 0.0F), (double)(yCoord + 0.0F), (double)this.zLevel).tex((double)((float)(minU + 0) * 0.00390625F), (double)((float)(minV + 0) * 0.00390625F)).endVertex();
+		bufferbuilder.pos(xCoord + 0.0F, yCoord + maxV, this.zLevel).tex((minU + 0) * 0.00390625F, (minV + maxV) * 0.00390625F).endVertex();
+		bufferbuilder.pos(xCoord + maxU, yCoord + maxV, this.zLevel).tex((minU + maxU) * 0.00390625F, (minV + maxV) * 0.00390625F).endVertex();
+		bufferbuilder.pos(xCoord + maxU, yCoord + 0.0F, this.zLevel).tex((minU + maxU) * 0.00390625F, (minV + 0) * 0.00390625F).endVertex();
+		bufferbuilder.pos(xCoord + 0.0F, yCoord + 0.0F, this.zLevel).tex((minU + 0) * 0.00390625F, (minV + 0) * 0.00390625F).endVertex();
 		tessellator.draw();
 	}
 
@@ -189,10 +188,10 @@ public abstract class GuiElement {
 		Tessellator tessellator = Tessellator.getInstance();
 		BufferBuilder bufferbuilder = tessellator.getBuffer();
 		bufferbuilder.begin(7, DefaultVertexFormats.POSITION_TEX);
-		bufferbuilder.pos((double)(xCoord + 0), (double)(yCoord + heightIn), (double)this.zLevel).tex((double)textureSprite.getMinU(), (double)textureSprite.getMaxV()).endVertex();
-		bufferbuilder.pos((double)(xCoord + widthIn), (double)(yCoord + heightIn), (double)this.zLevel).tex((double)textureSprite.getMaxU(), (double)textureSprite.getMaxV()).endVertex();
-		bufferbuilder.pos((double)(xCoord + widthIn), (double)(yCoord + 0), (double)this.zLevel).tex((double)textureSprite.getMaxU(), (double)textureSprite.getMinV()).endVertex();
-		bufferbuilder.pos((double)(xCoord + 0), (double)(yCoord + 0), (double)this.zLevel).tex((double)textureSprite.getMinU(), (double)textureSprite.getMinV()).endVertex();
+		bufferbuilder.pos(xCoord + 0, yCoord + heightIn, this.zLevel).tex(textureSprite.getMinU(), textureSprite.getMaxV()).endVertex();
+		bufferbuilder.pos(xCoord + widthIn, yCoord + heightIn, this.zLevel).tex(textureSprite.getMaxU(), textureSprite.getMaxV()).endVertex();
+		bufferbuilder.pos(xCoord + widthIn, yCoord + 0, this.zLevel).tex(textureSprite.getMaxU(), textureSprite.getMinV()).endVertex();
+		bufferbuilder.pos(xCoord + 0, yCoord + 0, this.zLevel).tex(textureSprite.getMinU(), textureSprite.getMinV()).endVertex();
 		tessellator.draw();
 	}
 
@@ -205,10 +204,10 @@ public abstract class GuiElement {
 		Tessellator tessellator = Tessellator.getInstance();
 		BufferBuilder bufferbuilder = tessellator.getBuffer();
 		bufferbuilder.begin(7, DefaultVertexFormats.POSITION_TEX);
-		bufferbuilder.pos((double)x, (double)(y + height), 0.0D).tex((double)(u * f), (double)((v + (float)height) * f1)).endVertex();
-		bufferbuilder.pos((double)(x + width), (double)(y + height), 0.0D).tex((double)((u + (float)width) * f), (double)((v + (float)height) * f1)).endVertex();
-		bufferbuilder.pos((double)(x + width), (double)y, 0.0D).tex((double)((u + (float)width) * f), (double)(v * f1)).endVertex();
-		bufferbuilder.pos((double)x, (double)y, 0.0D).tex((double)(u * f), (double)(v * f1)).endVertex();
+		bufferbuilder.pos(x, y + height, 0.0D).tex(u * f, (v + height) * f1).endVertex();
+		bufferbuilder.pos(x + width, y + height, 0.0D).tex((u + width) * f, (v + height) * f1).endVertex();
+		bufferbuilder.pos(x + width, y, 0.0D).tex((u + width) * f, v * f1).endVertex();
+		bufferbuilder.pos(x, y, 0.0D).tex(u * f, v * f1).endVertex();
 		tessellator.draw();
 	}
 
@@ -221,10 +220,10 @@ public abstract class GuiElement {
 		Tessellator tessellator = Tessellator.getInstance();
 		BufferBuilder bufferbuilder = tessellator.getBuffer();
 		bufferbuilder.begin(7, DefaultVertexFormats.POSITION_TEX);
-		bufferbuilder.pos((double)x, (double)(y + height), 0.0D).tex((double)(u * f), (double)((v + (float)vHeight) * f1)).endVertex();
-		bufferbuilder.pos((double)(x + width), (double)(y + height), 0.0D).tex((double)((u + (float)uWidth) * f), (double)((v + (float)vHeight) * f1)).endVertex();
-		bufferbuilder.pos((double)(x + width), (double)y, 0.0D).tex((double)((u + (float)uWidth) * f), (double)(v * f1)).endVertex();
-		bufferbuilder.pos((double)x, (double)y, 0.0D).tex((double)(u * f), (double)(v * f1)).endVertex();
+		bufferbuilder.pos(x, y + height, 0.0D).tex(u * f, (v + vHeight) * f1).endVertex();
+		bufferbuilder.pos(x + width, y + height, 0.0D).tex((u + uWidth) * f, (v + vHeight) * f1).endVertex();
+		bufferbuilder.pos(x + width, y, 0.0D).tex((u + uWidth) * f, v * f1).endVertex();
+		bufferbuilder.pos(x, y, 0.0D).tex(u * f, v * f1).endVertex();
 		tessellator.draw();
 	}
 }
